Proposed Updates
Great work on getting this started! I made some edits for readability and also wanted to clarify a few things.

Updated the description of TAS to keep it open for more languages
lowered the `i` on Instance because Tanzu Kubernetes Grid Instance is not a defined object. However, we do tend to refer to them as instances. 
The TKG management cluster does not provide cluster services but that is a function of TKG Extensions.